requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Audio processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1
audioread==3.0.1

# Machine Learning - flexible versions for Python 3.13
scikit-learn
numpy
scipy
joblib

# Text-to-Speech for synthetic data generation
gTTS==2.5.0
pyttsx3==2.90

# Utilities
requests==2.31.0

# Testing
pytest==7.4.3
httpx==0.25.2
